invoice-processor-web/requirements.txt
Django==5.1.3
python-decouple==3.8
django-bootstrap5==24.3
pdf2image==1.17.0
pytesseract==0.3.13
pdfplumber==0.11.4
requests==2.31.0
python-dotenv==1.0.0
openai==1.27.0
pillow==9.5.0